m2/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
tabulate==0.9.0
python-multipart==0.0.6
pydantic==2.5.3
fastapi
pandas
psycopg2-binary
pydantic
python-multipart
sqlalchemy
tabulate
uvicorn
bcrypt
passlib
python-jose
